+++ pages/category.js
@@ -1,20 +1,22 @@
 import Head from 'next/head'
 import { Box } from '@chakra-ui/react'
-import Navbar from '@/components/Navbar'
+import Navigation from '@/components/pages/category'
+import Products from '@/components/pages/category/products'
 import Footer from '@/components/Footer'
 import LandingContents from '@/components/pages/landing'
 
-const Home = () => {
+const Category = () => {
     return (
         <Box>
             <Head>
                 <title>Swift Shop | Home</title>
             </Head>
-            <Navbar />
-            <LandingContents />
+            <Navigation />
+            <Products />
+            
             <Footer />
         </Box>
     )
 }
 
-export default Home
+export default Category
