+++ pages/index.js
@@ -1,22 +1,20 @@
 import Head from 'next/head'
 import { Box } from '@chakra-ui/react'
-import Navigation from '@/components/pages/category'
-import Products from '@/components/pages/category/products'
+import Navbar from '@/components/Navbar'
 import Footer from '@/components/Footer'
 import LandingContents from '@/components/pages/landing'
 
-const Category = () => {
+const Home = () => {
     return (
         <Box>
             <Head>
                 <title>Swift Shop | Home</title>
             </Head>
-            <Navigation />
-            <Products />
-            
+            <Navbar />
+            <LandingContents />
             <Footer />
         </Box>
     )
 }
 
-export default Category
+export default Home
